pyproject.toml
[tool.poetry]
name = "code-rag-agent"
version = "0.1.0"
description = "Hybrid Graph and Vector RAG for software repository analysis"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
llama-index = "^0.10.43"
pygithub = "^2.3.0"
neo4j = "^5.21.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
